+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mmunoz-f <mmunoz-f@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/22 20:04:28 by marvin            #+#    #+#             */
-/*   Updated: 2021/03/10 12:59:38 by mmunoz-f         ###   ########.fr       */
+/*   Updated: 2022/01/10 18:14:21 by mmunoz-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] == s2[i] && i < n && s1[i])
