+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: mmunoz-f <mmunoz-f@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/22 19:55:34 by marvin            #+#    #+#             */
-/*   Updated: 2021/02/01 17:30:12 by mmunoz-f         ###   ########.fr       */
+/*   Updated: 2022/01/10 18:14:09 by mmunoz-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(char *big, const char *little, size_t len)
 {
-	unsigned int	i;
-	unsigned int	total;
+	size_t	i;
+	size_t	total;
 
 	i = 0;
 	if (!(ft_strlen(little)))
